Get read of AbstractTestWorkspaceProvider; remake WorkspaceTemplate to enum
