+++ selenium/che-selenium-core/src/test/java/org/eclipse/che/selenium/core/utils/WorkspaceDtoDeserializerTest.java
@@ -13,6 +13,7 @@ package org.eclipse.che.selenium.core.utils;
 
 import java.lang.reflect.Field;
 import org.eclipse.che.selenium.core.constant.Infrastructure;
+import org.eclipse.che.selenium.core.workspace.WorkspaceTemplate;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
@@ -32,15 +33,7 @@ public class WorkspaceDtoDeserializerTest {
   @Test
   public void shouldBeAbleToGetWorkspaceConfigFromResource() {
 
-    Assert.assertNotNull(deserializer.deserializeWorkspaceTemplate("default.json"));
-  }
-
-  @Test(
-      expectedExceptions = RuntimeException.class,
-      expectedExceptionsMessageRegExp =
-          "resource /templates/workspace/openshift/some.json relative to org.eclipse.che.selenium.core.utils.WorkspaceDtoDeserializer not found.")
-  public void shouldFailIfResourceIsNotFound() {
-    deserializer.deserializeWorkspaceTemplate("some.json");
+    Assert.assertNotNull(deserializer.deserializeWorkspaceTemplate(WorkspaceTemplate.DEFAULT));
   }
 
   @Test(
@@ -48,6 +41,6 @@ public class WorkspaceDtoDeserializerTest {
       expectedExceptionsMessageRegExp =
           "com.google.gson.stream.MalformedJsonException: Expected ':' at line 3 column 8 path \\$\\.werwerw")
   public void shouldFailIfNotAJson() {
-    deserializer.deserializeWorkspaceTemplate("notAJson.json");
+    deserializer.deserializeWorkspaceTemplate(WorkspaceTemplate.BROKEN);
   }
 }
