--- selenium/che-selenium-core/src/test/resources/templates/workspace/openshift/notAJson.json
@@ -1,4 +1,0 @@
-{
-  werwerw
-      awerw
-}
