+++ selenium/che-selenium-test/src/test/java/org/eclipse/che/selenium/gwt/CheckSimpleGwtAppTest.java
@@ -20,7 +20,6 @@ import static org.eclipse.che.selenium.core.constant.TestTimeoutsConstants.REDRA
 import static org.eclipse.che.selenium.core.workspace.WorkspaceTemplate.UBUNTU_JDK8;
 
 import com.google.inject.Inject;
-import com.google.inject.name.Named;
 import java.net.URL;
 import java.nio.file.Paths;
 import org.eclipse.che.api.workspace.shared.dto.ServerConfigDto;
@@ -34,9 +33,8 @@ import org.eclipse.che.selenium.core.project.ProjectTemplates;
 import org.eclipse.che.selenium.core.user.DefaultTestUser;
 import org.eclipse.che.selenium.core.utils.WaitUtils;
 import org.eclipse.che.selenium.core.utils.WorkspaceDtoDeserializer;
+import org.eclipse.che.selenium.core.workspace.CheTestWorkspaceProvider;
 import org.eclipse.che.selenium.core.workspace.TestWorkspace;
-import org.eclipse.che.selenium.core.workspace.TestWorkspaceImpl;
-import org.eclipse.che.selenium.core.workspace.TestWorkspaceLogsReader;
 import org.eclipse.che.selenium.pageobject.Consoles;
 import org.eclipse.che.selenium.pageobject.Ide;
 import org.eclipse.che.selenium.pageobject.ProjectExplorer;
@@ -65,13 +63,9 @@ public class CheckSimpleGwtAppTest {
   @Inject private TestWorkspaceServiceClient workspaceServiceClient;
   @Inject private TestProjectServiceClient testProjectServiceClient;
   @Inject private WorkspaceDtoDeserializer workspaceDtoDeserializer;
-  @Inject private TestWorkspaceLogsReader testWorkspaceLogsReader;
   @Inject private DefaultTestUser testUser;
   @Inject private SeleniumWebDriver seleniumWebDriver;
-
-  @Inject
-  @Named("tests.workspacelogs_dir")
-  private String workspaceLogsDir;
+  @Inject private CheTestWorkspaceProvider testWorkspaceProvider;
 
   private String projectName;
 
@@ -93,15 +87,8 @@ public class CheckSimpleGwtAppTest {
             newDto(ServerConfigDto.class).withProtocol("http").withPort("9876"));
 
     testWorkspace =
-        new TestWorkspaceImpl(
-            NameGenerator.generate("check-gwt-test", 4),
-            testUser,
-            4,
-            true,
-            workspace,
-            workspaceServiceClient,
-            testWorkspaceLogsReader,
-            workspaceLogsDir);
+        testWorkspaceProvider.createWorkspace(
+            NameGenerator.generate("check-gwt-test", 4), testUser, 4, true, workspace);
 
     URL resource = getClass().getResource("/projects/web-gwt-java-simple");
     testProjectServiceClient.importProject(
